requirements.txt
aiofiles
aiogram
aiohttp[speedups]
aioredis
alphabet_detector
beautifulsoup4==4.9.1
better_profanity
bing_image_downloader
bleach
bs4
cachetools
cloudscraper
countryinfo
cowpy
cryptg
cryptography>=3.3.2
dateparser==1.1.1
dnspython
emoji==1.7.0
envparse
faker
feedparser
ffmpeg-python
fontTools
futures
geopy
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
glitch_this
glitch_this
google-api-python-client
google-trans-new
google-trans-new
googletrans
gpytranslate
gtts
hachoir
heroku3
hiredis
html-telegraph-poster
html2text
httpx
humanize
img2pdf
jikanpy
launchpadlib==1.10.13
lxml
lyricsgenius
markdown
markdown2>=2.4.0
motor<2.4.0,>=2.1.0
multicolorcaptcha
nekos.py
nltk
nudepy
NumPy
odmantic
opencv-python-headless
pendulum
Pillow
pretty_errors
psutil
psycopg2-binary
PyDictionary==2.0.1
pydub
pykeyboard
pymongo==3.12.3
pynewtonmath
pyrate-limiter
pyrogram==2.0.52
python-arq
python-dotenv
python-rapidjson
python-telegram-bot==13.11
pytimeparse
pytube
pytz
pyYAML>=5.1.2
pySmartDL
redis
regex
requests
requests_html
safoneapi==1.0.20
search_engine_parser
secureme
selenium
spamwatch
speedtest-cli
spongemock
sqlalchemy==1.3.20
telegraph
telethon==1.24.0
textblob
tgcrypto
tswift
ujson
uvloop
wget
wikipedia
youtube_dl
youtube_search
youtube_search_python
yt-dlp
zalgo-text